requirements.txt
Flask==3.1.0
gunicorn==20.1.0
torch==1.13.1
torchvision==0.14.1
Pillow==11.0.0
